Add UnitMeasure and Product models
