+++ chalicelib/controllers/unit_measure_controller.py
@@ -1,4 +1,4 @@
-from chalicelib.models.unit_measure_model import UnitMeasure, Session
+from chalicelib.models.models import UnitMeasure, Session
 
 
 def create_unit_measure(name: str) -> dict and int:
